third_party/blink/web_tests/virtual/cssom-getcomputedstyle-pseudo-no-colon-disabled/external/wpt/css/cssom/getComputedStyle-pseudo-expected.txt

This is a testharness.js-based test.
[FAIL] Resolution of width is correct when pseudo-element argument is ignored (due to no colon)
  assert_equals: expected "100px" but got "50px"
[FAIL] Unknown pseudo-element with a known identifier: file-selector-button
  assert_equals: Should return the ::file-selector-button style expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
[FAIL] Unknown pseudo-element with a known identifier: placeholder
  assert_equals: Should return the ::placeholder style expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
